pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "bread-ai-fiscal"
version = "0.1.0"
description = ""
authors = [
    {name = "João Vitor",email = "joaovitoralcantara30@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "agno (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "google-genai (>=1.47.0,<2.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
